--- src/main/java/com/example/seoulpublicdata2025backend/domain/consumption/service/MemberConsumptionService.java
@@ -1,7 +1,0 @@
-package com.example.seoulpublicdata2025backend.domain.consumption.service;
-
-import com.example.seoulpublicdata2025backend.domain.company.dto.CompanyLocationTypeDto;
-
-public interface MemberConsumptionService {
-    void saveConsumption(CompanyLocationTypeDto companyDto, Long totalPrice);
-}
