--- src/main/java/com/example/seoulpublicdata2025backend/domain/kakaoSocialLogin/controller/MemberController.java
@@ -1,27 +1,0 @@
-package com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.controller;
-
-import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.dto.SignupRequestDto;
-import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.dto.SignupResponseDto;
-import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.entity.Member;
-import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.service.MemberService;
-import com.example.seoulpublicdata2025backend.global.swagger.annotations.member.SignUpDocs;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/member")
-public class MemberController {
-
-    private final MemberService memberService;
-
-    @PostMapping("/signup")
-    @SignUpDocs
-    public ResponseEntity<SignupResponseDto> signup(@Valid @RequestBody SignupRequestDto dto) {
-        SignupResponseDto response = memberService.updateMember(dto);
-        return ResponseEntity.ok(response);
-    }
-}
