+++ src/main/java/com/example/seoulpublicdata2025backend/domain/member/controller/SignupPageController.java
@@ -1,4 +1,4 @@
-package com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.controller;
+package com.example.seoulpublicdata2025backend.domain.member.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
