+++ src/main/java/com/example/seoulpublicdata2025backend/domain/member/dto/MemberConsumptionResponseDto.java
@@ -1,0 +1,12 @@
+package com.example.seoulpublicdata2025backend.domain.member.dto;
+
+import com.example.seoulpublicdata2025backend.domain.company.entity.CompanyType;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class MemberConsumptionResponseDto {
+    private CompanyType companyType;
+    private Long totalPrice;
+}
