+++ src/main/java/com/example/seoulpublicdata2025backend/domain/member/dto/SignupResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.dto;
+package com.example.seoulpublicdata2025backend.domain.member.dto;
 
 import lombok.Getter;
