+++ src/main/java/com/example/seoulpublicdata2025backend/domain/member/service/MemberConsumptionService.java
@@ -1,0 +1,14 @@
+package com.example.seoulpublicdata2025backend.domain.member.service;
+
+import com.example.seoulpublicdata2025backend.domain.company.dto.CompanyLocationTypeDto;
+import com.example.seoulpublicdata2025backend.domain.company.entity.CompanyType;
+import com.example.seoulpublicdata2025backend.domain.member.dto.MemberConsumptionResponseDto;
+import java.util.List;
+
+public interface MemberConsumptionService {
+    void saveConsumption(CompanyLocationTypeDto companyDto, Long totalPrice);
+
+    List<MemberConsumptionResponseDto> findConsumptionByMember();
+
+    List<MemberConsumptionResponseDto> findConsumptionByMemberAndCompanyType(CompanyType companyType);
+}
