+++ src/main/java/com/example/seoulpublicdata2025backend/domain/review/dto/CompanyReviewDto.java
@@ -1,7 +1,7 @@
 package com.example.seoulpublicdata2025backend.domain.review.dto;
 
 import com.example.seoulpublicdata2025backend.domain.company.entity.Company;
-import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.entity.Member;
+import com.example.seoulpublicdata2025backend.domain.member.entity.Member;
 import com.example.seoulpublicdata2025backend.domain.review.entity.ReviewCategory;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
