+++ src/main/java/com/example/seoulpublicdata2025backend/domain/review/service/ReviewServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.seoulpublicdata2025backend.domain.review.service;
 
 import com.example.seoulpublicdata2025backend.domain.company.entity.Company;
-import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.entity.Member;
+import com.example.seoulpublicdata2025backend.domain.member.entity.Member;
 import com.example.seoulpublicdata2025backend.domain.review.dao.CompanyReviewRepository;
 import com.example.seoulpublicdata2025backend.domain.review.dto.CompanyReviewDto;
 import com.example.seoulpublicdata2025backend.domain.review.dto.ReviewDto;
