+++ src/main/java/com/example/seoulpublicdata2025backend/global/swagger/annotations/member/KakaoLoginCheckDocs.java
@@ -1,11 +1,9 @@
 package com.example.seoulpublicdata2025backend.global.swagger.annotations.member;
 
-import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.dto.KakaoAuthResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
-import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
