+++ src/test/java/com/example/seoulpublicdata2025backend/domain/review/controller/ReviewControllerTest.java
@@ -1,7 +1,7 @@
 package com.example.seoulpublicdata2025backend.domain.review.controller;
 
 import com.example.seoulpublicdata2025backend.domain.company.entity.Company;
-import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.entity.Member;
+import com.example.seoulpublicdata2025backend.domain.member.entity.Member;
 import com.example.seoulpublicdata2025backend.domain.review.dto.CompanyReviewDto;
 import com.example.seoulpublicdata2025backend.domain.review.dto.ReviewDto;
 import com.example.seoulpublicdata2025backend.domain.review.entity.ReviewCategory;
